使用有符号整数比较当前版面最大在线人数
git-svn-id: http://svn.kcn.cn/repos/kbs/trunk@11584 de4889a5-8ad4-4bf3-bd87-57997c21d515

+++ kbs_bbs/libBBS/bcache.c
@@ -575,7 +575,7 @@ void board_setcurrentuser(int idx,int num)
 #endif
 #ifdef RECORDMAXONLINE
         /* save board max online record */
-        if (brdshm->bstatus[idx - 1].currentusers > brdshm->bstatus[idx - 1].maxonline) {
+        if ((int)brdshm->bstatus[idx - 1].currentusers > (int)brdshm->bstatus[idx - 1].maxonline) {
             brdshm->bstatus[idx - 1].maxonline = brdshm->bstatus[idx - 1].currentusers;
             brdshm->bstatus[idx - 1].maxtime = time(0);
         }
